fix: clippy + fmt

+++ src/grc/converter/dsb.rs
@@ -1,7 +1,7 @@
 use super::super::converter_helper::{BlockConverter, ConnectorAdapter, DefaultPortAdapter};
 use super::{BlockInstance, Grc2FutureSdr};
-use anyhow::Result;
 use crate::blocks::dsb_fc;
+use anyhow::Result;
 use futuresdr::runtime::Flowgraph;
 
 pub struct DsbConverter {}
@@ -12,7 +12,7 @@ impl BlockConverter for DsbConverter {
         blk: &BlockInstance,
         fg: &mut Flowgraph,
     ) -> Result<Box<dyn ConnectorAdapter>> {
-        let q_value = Grc2FutureSdr::parameter_as_f32(blk, "q_value", "0.0")? as f32;
+        let q_value = Grc2FutureSdr::parameter_as_f32(blk, "q_value", "0.0")?;
         let blk = dsb_fc(q_value);
         let blk = fg.add_block(blk);
         let blk = DefaultPortAdapter::new(blk.into());
@@ -20,4 +20,3 @@ impl BlockConverter for DsbConverter {
         Ok(blk)
     }
 }
-
